Make isIntersectingViewport work with SVG elements (#2586)

+++ lib/PuppeteerSharp/IElementHandle.cs
@@ -119,11 +119,11 @@ namespace PuppeteerSharp
         Task HoverAsync();
 
         /// <summary>
-        /// Evaluates if the element is visible in the current viewport.
+        /// Evaluates if the element is visible in the current viewport. If an element is an SVG, we check if the svg owner element is in the viewport instead.
         /// </summary>
         /// <param name="threshold">A number between 0 and 1 specifying the fraction of the element's area that must be visible to pass the check.</param>
         /// <returns>A task which resolves to true if the element is visible in the current viewport.</returns>
-        Task<bool> IsIntersectingViewportAsync(int threshold = 0);
+        Task<bool> IsIntersectingViewportAsync(decimal threshold = 0);
 
         /// <summary>
         /// Focuses the element, and then uses <see cref="IKeyboard.DownAsync(string, DownOptions)"/> and <see cref="IKeyboard.UpAsync(string)"/>.
